Separated packages for go routines
